Add a --sync mode that lets a user download pre-uploaded article tracking IDs
In my case, I already have a load of article published on my blog.  They
weren't uploaded with gitblogger so gitblogger's notes branch doesn't
have a copy of the post ID.

This patch adds a --sync mode to let a user bootstrap their repository
in just such a situation.  The title of the post is used as the key to
match the local article with the blog-posted article.

It is making the assumption that the posted article matches the local
article -- it makes no attempt to upload every local article again.
This would fail anyway as Google set a limit on the number of posts per
day.

Also, modification mode doesn't work properly yet, so it would only
wreck previously posted articles.