Add a mode that allows an existing set of articles to bootstrap a whole blog via an import file
Rather than adding the posts itself, this new --bootstrap mode allows a
caller to create an XML file that can be fed to blogger's "import blog"
function.

In principle, if a user deleted their entire blog, they could use this
mode to reconstruct every post in one go.  It can also be used for a
repository that already contains a set of articles, and to bring a blog
up to date without tripping blogger's 'posts-per-day' limit.

After importing that bootstrap file to blogger, the user should run the
--sync mode to download all of the post IDs.  That relies on all the
articles having different titles though.