ikiwikiToAtom() now returns a utf-8 converted version of the atom XML
Without this, the unicode that was returned would be handed to httplib
for transmission, which tried to encode it as ASCII -- which would be
impossible and would crash the transmission.

This fix makes sure that the data is already encoded as something that
can be safely transmitted by httplib (and matches the encoding of the
Atom XML).