Make article modification work, content only at this stage, meta information unchanged
The trick is to construct a new XML DOM object containing the new
article content node, then extract that node and push it into the
existing article DOM object using minidom.importNode().

Then the original content node can be replaced successfully.

Previously the replacement was being done by turning the new article
content into a text node -- that didn't work because that escaped all of
the HTML tags, so the article was replaced with a verbatim, literal
rendering of the HTML.