Fix incorrect timezone being sent as published date
I'm not entirely sure what the datetime module is for in python, but it
is unwilling or unable to handle times with timezones without you first
knowing the timezone the program is running in.  That's never going to
happen, we can't hard code timezones and adding a specific method of
setting timezone to the config file is crazy.

Instead, I've found that python provides wrappers around the standard
libc time handling functions, which are perfectly sufficient for what we
need.  Now, the date field from the article file is parsed like this:

 localtime_tuple = time.strptime( ... )
 absolute_epoch = time.mktime( localtime_tuple )
 utc_tuple = time.gmtime( absolute_epoch )
 utc_publish = time.strftime('%Y-%m-%dT%H:%M:%S+0000, utc_tuple)

So, strptime() returns a local timezone tuple, which is converted to
epoch time by mktime, which is converted to a UTC tuple using gmtime(),
which is formatted with the "+0000" suffix needed to describe it as UTC
for blogger.