Fix utf-8 problems by making unicode throughout, converting after article read
The key changes are

  open( filename, mode='rb', encoding='utf-8' )

Which automatically makes subsequent reads from this file convert to
unicode from utf-8.  Similarly, when reading articles from git...

  md_source = subprocess.Popen(["git", "cat-file", "-p", \
               article[0]], stdout=subprocess.PIPE).communicate()[0]
  md_source = unicode( md_source, 'utf-8' )

This ensures that all data is processed internally as unicode, and then
for output python's "print" seems to take care of itself, and for
httplib, we simply encode to 'utf-8' before transmitting:

  response, content = self.http.request( postURL, 'PUT',
               headers=headers, body=upload.encode('utf-8') )

This seems to have solved the character set problems for now.  I'm still
not entirely convinced that python print is going to do the correct
thing, as there are a lot of complains on the web about not being able
to set the encoding for stdout.